change name file formRegister for registerForm, chance structure Form component and whote on registerForm.js
